+++ tests/test_widgets/test_mining_schema.py
@@ -61,7 +61,7 @@ def test_df(mining_schema_df):
 
     # Test ignoring unknown columns
     schema_df_new = mining_schema_df.drop_duplicates().copy()
-    schema_df_new["unknown_column"] = [i for i in range(len(schema_df_new))]
+    schema_df_new["unknown_column"] = list(range(len(schema_df_new)))
     mining_schema = MiningSchema()
     with pytest.warns(UserWarning, match=r"column.* unknown_column"):
         mining_schema.add_from_df(schema_df_new)